+++ CHANGELOG.md
@@ -4,4 +4,5 @@ Version 0.0.1 (WIP)
 - Cleaned out obsolete local version control stuff
 - Created Add/Edit/Clear functions
 - Expanded part catalog for scaling
-- Board fixed at (0,0,0) 
+- Board fixed at (0,0,0)
+- Added part hover scaling